refactor(ctrls): use new names for service entry
